代表者挨拶を会社概要ページに移設
